flatten file structure and eliminate circular imports
