+++ samples/chaincode/badge/src/BadgeInstantiationFactory.ts
@@ -7,11 +7,8 @@ import { PrimedComponentFactory } from "@microsoft/fluid-aqueduct";
 import { SharedCell } from "@microsoft/fluid-cell";
 import { SharedMap } from "@microsoft/fluid-map";
 import { SharedObjectSequence } from "@microsoft/fluid-sequence";
-import { Badge } from "./index";
+import { Badge } from "./Badge";
 
-/**
- * This is where you define all your Distributed Data Structures
- */
 export const BadgeInstantiationFactory = new PrimedComponentFactory(
     Badge,
     [
