+++ samples/chaincode/badge/src/IBadgeType.ts
@@ -1,0 +1,12 @@
+/*!
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+import { IIconProps } from "office-ui-fabric-react/lib/Icon";
+
+export interface IBadgeType {
+    key: string;
+    text: string;
+    iconProps: IIconProps;
+}
